Move file to correct place in structure
