+++ src/features/events/components/EventParticipantsCard.tsx
