+++ src/pages/organize/[orgId]/projects/[campId]/events/[eventId]/index.tsx
@@ -6,7 +6,7 @@ import { scaffold } from 'utils/next';
 import EventDataModel from 'features/events/models/EventDataModel';
 import EventLayout from 'features/events/layout/EventLayout';
 import EventOverviewCard from 'features/events/components/EventOverviewCard';
-import EventParticipantsCard from 'features/events/components/EventOverviewCard/EventParticipantsCard';
+import EventParticipantsCard from 'features/events/components/EventParticipantsCard';
 import LocationsModel from 'features/events/models/LocationsModel';
 import useModel from 'core/useModel';
 import ZUIFuture from 'zui/ZUIFuture';
